refactor: Separate sheet UI utils and data sheet utils
